Move softDeleteMiddleware into its own folder
